Remove socketio, eventlet
