+++ .env/.dev
@@ -4,4 +4,3 @@ DATABASE_URL=postgresql://flask_celery:flask_celery@db/flask_celery
 SECRET_KEY=not_so_secret
 CELERY_BROKER_URL=redis://redis:6379/0
 CELERY_RESULT_BACKEND=redis://redis:6379/0
-SOCKETIO_MESSAGE_QUEUE=redis://redis:6379/0
