+++ compose/local/flask/start
@@ -5,5 +5,4 @@ set -o pipefail
 set -o nounset
 
 flask db upgrade
-# flask run --host=0.0.0.0
-python app.py
+flask run --host=0.0.0.0
