+++ cv_model_serve/celery_utils.py
