+++ cv_model_serve/image_classifier/__init__.py
@@ -1,1 +1,1 @@
-from . import events, tasks  # noqa
+from . import tasks  # noqa
