+++ pyproject.toml
@@ -10,14 +10,12 @@ Flask = "^2.1.2"
 celery = "^5.2.7"
 redis = "^4.3.4"
 flower = "^1.1.0"
-eventlet = "^0.33.1"
 requests = "^2.28.1"
 Flask-Migrate = "^3.1.0"
 Flask-SQLAlchemy = "^2.5.1"
 Flask-CeleryExt = "^0.4.3"
 psycopg2-binary = "^2.9.3"
 watchgod = "^0.8.2"
-Flask-SocketIO = "^5.2.0"
 numpy = "^1.23.1"
 Pillow = "^9.2.0"
 tensorflow = "^2.9.1"
@@ -25,6 +23,7 @@ tensorflow = "^2.9.1"
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
 black = {version = "^22.6.0", allow-prereleases = true}
+isort = "^5.10.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
